+++ Src/SLib/ssvg.cpp
@@ -515,7 +515,7 @@ int SSvg::_GetPoints(SStrScan & rScan, const char * pTxt, FloatArray & rList)
 	rScan.Skip();
 	while(rScan.GetDotPrefixedNumber(temp_buf)) {
 		// @v10.4.10 float val = temp_buf.ToFloat();
-		float val = atof(temp_buf); // @v10.4.10 
+		float val = static_cast<float>(atof(temp_buf)); // @v10.4.10 
 		float temp_val = temp_buf.ToFloat();
 		assert(val == temp_val);
 		rList.add(val);
